apps/api/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2

# Auth
python-jose[cryptography]==3.3.0
httpx==0.27.2

# Celery + Redis
celery[redis]==5.4.0
redis==5.1.1

# Supabase Storage
supabase==2.7.2
storage3==0.7.7

# Data processing (shared with pipeline)
pandas==2.2.3
numpy==1.26.4

# NLP (for pipeline core modules)
nltk==3.9.1

# OpenAI (for index generation + resume optimization)
openai==1.12.0

# Pydantic
pydantic==2.10.3
pydantic-settings==2.5.2

# PDF parsing (resume upload worker)
pymupdf==1.24.11
pdfplumber==0.11.4

# PDF generation (optimized resume export)
fpdf2==2.8.1

# Browser automation (pipeline scraping)
selenium==4.25.0

# Utilities
python-dotenv==1.0.1
cryptography==43.0.1

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
aiosqlite>=0.20.0
